chatbot_hugging/requirements_huggingface.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.20.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
pandas>=2.0.0
numpy>=1.24.0
